Lots of changes - multiple lists now works I think
Finished the ability to run multiple lists. Every file in the skuLists directory will be used one at a time, and will generate a results page in the results directory that has a similar name to the original sku list file. This means that the results and skuLists directories are now baked into the program. Updated the todo list. Added a favicon (kinda cheesy) and included the GIMP files. Re-organized everything and created a resources and inProgress directory to keep the home directory clean as the list of files I'm working on continues to grow